Changing int64s to ints

+++ take.go
@@ -7,15 +7,15 @@ package pipe
 // Accept only the given number of items from the input pipe. After that number
 // has been received, all input messages will be ignored and the output channel
 // will be closed.
-func (p *Pipe) Take(num int64) *Pipe {
+func (p *Pipe) Take(num int) *Pipe {
 	p.addStage()
 	go p.takerHandler(num, p.length-1)()
 
 	return p
 }
 
-func (p *Pipe) takerHandler(num int64, pos int) func() {
-	var count int64
+func (p *Pipe) takerHandler(num int, pos int) func() {
+	var count int
 	return func() {
 		// only send num items
 		for count = 0; count < num; count++ {
